:truck: making error messages more generic and renaming imports
