+++ src/utils/errorMessages.ts
@@ -23,12 +23,12 @@ export const missingInvalidTokenError = () => {
   };
 };
 
-export const missingUserError = () => {
+export const missingDocError = (docName: string) => {
   return {
     errors: [
       {
-        type: "user error",
-        message: "User doesn't exist.",
+        type: `${docName} error`,
+        message: `${docName} doesn't exist.`,
       },
     ],
   };
